Extract v2flags logic
